Update as-path-prepend data format
