+++ lib/netomox/version.rb
@@ -2,5 +2,5 @@
 
 module Netomox
   # Netomox version
-  VERSION = '0.10.0.pre1'
+  VERSION = '0.10.0.pre2'
 end
